+++ share/src/main/resources/alfresco/site-webscripts/com/parashift/onlyoffice-edit.get.js
@@ -14,7 +14,7 @@ if (url.args.sample) query +="&sample=" + url.args.sample;
     model.onlyofficeUrl = pObj.onlyofficeUrl;
     model.editorConfig = JSON.stringify(pObj.editorConfig);
     model.docTitle = pObj.editorConfig.document.title;
-    model.documentType = pObj.editorConfig.document.fileType == 'docxf' || pObj.editorConfig.document.fileType == 'oform' ? pObj.editorConfig.document.fileType : pObj.editorConfig.documentType;
+    model.documentType = pObj.editorConfig.document.fileType == 'docxf' ? pObj.editorConfig.document.fileType : pObj.editorConfig.documentType;
     model.folderNode = pObj.folderNode;
     model.demo = pObj.demo;
     model.favorite = pObj.favorite;