+++ CHANGELOG.md
@@ -6,9 +6,12 @@
 - document server v6.4 and earlier is no longer supported
 - setting authorization header on settings page
 - docs cloud banner on settings page
+- filling pdf
 
 ## Changed
 - Alfresco v6.* and earlier is no longer supported
+- default conversion format (from docxf to pdf instead oform)
+- remove filling for oform
 
 ## 6.1.0
 ## Added